Rename file realdataconvenience.h -> realdata_types.h for consistency
