+++ tests/testdareflcore/realdatamodel.test.cpp
@@ -9,7 +9,7 @@
 
 #include "google_test.h"
 
-#include <darefl/model/datasetconvenience.h>
+#include <darefl/model/realdata_types.h>
 #include <darefl/model/realdataitems.h>
 #include <darefl/model/realdatamodel.h>
